Refactor [skip ci]
